Refactor data output configurations to separate folder
