+++ backend/app/data_output_configuration/redshift/model.py
@@ -1,9 +1,9 @@
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.data_outputs.base_model import BaseDataOutputConfiguration
+from app.data_output_configuration.base_model import BaseDataOutputConfiguration
 
 
-class SnowflakeDataOutput(BaseDataOutputConfiguration):
+class RedshiftDataOutput(BaseDataOutputConfiguration):
     database: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
     schema: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
     table: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
@@ -14,6 +14,6 @@ class SnowflakeDataOutput(BaseDataOutputConfiguration):
     table_path: Mapped[str] = mapped_column(nullable=True, use_existing_column=True)
 
     __mapper_args__ = {
-        "polymorphic_identity": "SnowflakeDataOutput",
+        "polymorphic_identity": "RedshiftDataOutput",
         "polymorphic_load": "inline",
     }
