+++ backend/app/data_output_configuration/schema_union.py
@@ -2,12 +2,12 @@ from typing import Annotated, Union
 
 from pydantic import Field
 
-from app.data_outputs.data_output_types import DataOutputTypes
-from app.data_outputs.databricks_data_output.schema import DatabricksDataOutput
-from app.data_outputs.glue_data_output.schema import GlueDataOutput
-from app.data_outputs.redshift_data_output.schema import RedshiftDataOutput
-from app.data_outputs.s3_data_output.schema import S3DataOutput
-from app.data_outputs.snowflake_data_output.schema import SnowflakeDataOutput
+from app.data_output_configuration.data_output_types import DataOutputTypes
+from app.data_output_configuration.databricks.schema import DatabricksDataOutput
+from app.data_output_configuration.glue.schema import GlueDataOutput
+from app.data_output_configuration.redshift.schema import RedshiftDataOutput
+from app.data_output_configuration.s3.schema import S3DataOutput
+from app.data_output_configuration.snowflake.schema import SnowflakeDataOutput
 
 DataOutputs = Union[
     S3DataOutput,
